backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0

# Deep Learning & Computer Vision
tensorflow>=2.16.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Face Detection
mtcnn>=0.1.0
facenet-pytorch>=2.5.0

# Video Processing
moviepy>=1.0.0
imageio>=2.33.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
